+++ wcomponents-core/src/test/java/com/github/bordertech/wcomponents/WebUtilities_Test.java
@@ -1,19 +1,19 @@
 package com.github.bordertech.wcomponents;
 
+import com.github.bordertech.wcomponents.container.TransformXMLTestHelper;
+import com.github.bordertech.wcomponents.util.Config;
+import com.github.bordertech.wcomponents.util.ConfigurationProperties;
+import com.github.bordertech.wcomponents.util.SystemException;
+import com.github.bordertech.wcomponents.util.mock.MockRequest;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import junit.framework.Assert;
-import org.apache.commons.configuration.Configuration;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
-
-import com.github.bordertech.wcomponents.util.Config;
-import com.github.bordertech.wcomponents.util.ConfigurationProperties;
-import com.github.bordertech.wcomponents.util.SystemException;
 
 /**
  * WebUtilities_Test - unit tests for {@link WebUtilities}.
@@ -23,19 +23,19 @@ import com.github.bordertech.wcomponents.util.SystemException;
  */
 public class WebUtilities_Test extends AbstractWComponentTestCase {
 
-	private static Configuration originalConfig;
-
 	@BeforeClass
 	public static void setUp() {
-		originalConfig = Config.getInstance();
-		Configuration copyConfig = Config.copyConfiguration(originalConfig);
-		Config.setConfiguration(copyConfig);
+		Config.getInstance().setProperty(ConfigurationProperties.THEME_CONTENT_PATH, "");
+		TransformXMLTestHelper.reloadTransformer();
 	}
 
+	/**
+	 * When these tests are done put things back as they were.
+	 */
 	@AfterClass
-	public static void tearDown() {
-		// Remove overrides
-		Config.setConfiguration(originalConfig);
+	public static void tearDownClass() {
+		Config.reset();
+		TransformXMLTestHelper.reloadTransformer();
 	}
 
 	@Test
@@ -102,6 +102,20 @@ public class WebUtilities_Test extends AbstractWComponentTestCase {
 				WComponent.class, root));
 		Assert.assertNull("Incorrect ancestor returned", WebUtilities.getClosestOfClass(
 				WButton.class, dropdown));
+	}
+
+	@Test
+	public void testGetTop() {
+		WContainer root = new WContainer();
+		WTabSet tabs = new WTabSet();
+		WDropdown dropdown = new WDropdown();
+
+		root.add(tabs);
+		tabs.addTab(dropdown, "dropdown tab", WTabSet.TAB_MODE_CLIENT);
+
+		Assert.assertEquals("Incorrect top component returned for child", root, WebUtilities.getTop(dropdown));
+		Assert.assertEquals("Incorrect top component returned for child", root, WebUtilities.getTop(tabs));
+		Assert.assertEquals("Incorrect top component returned for top", root, WebUtilities.getTop(root));
 	}
 
 	// @Test
@@ -537,6 +551,40 @@ public class WebUtilities_Test extends AbstractWComponentTestCase {
 				WebUtilities.getParentNamingContext(child2));
 		Assert.assertEquals("Naming context for child3 should be context3", context3,
 				WebUtilities.getParentNamingContext(child3));
+	}
+
+	@Test
+	public void testRenderWithPlainText() {
+		String msg = "Test error message";
+		WText text = new WText(msg);
+		String output = WebUtilities.render(text);
+		Assert.assertEquals("Invalid output returned", msg, output);
+	}
+
+	@Test
+	public void testRenderWithXML() {
+		WText text = new WText(TransformXMLTestHelper.TEST_XML);
+		text.setEncodeText(false);
+		String output = WebUtilities.render(text);
+		Assert.assertEquals("Invalid output with XML", TransformXMLTestHelper.TEST_XML, output);
+	}
+
+	@Test(expected = SystemException.class)
+	public void testRenderToHtmlWithPlainText() {
+		String msg = "Test error message";
+		WText text = new WText(msg);
+		// Text will fail as it is not valid XML
+		String output = WebUtilities.renderWithTransformToHTML(new MockRequest(), text, false);
+		Assert.assertEquals("Invalid html output returned", msg, output);
+	}
+
+	@Test
+	public void testRenderToHtmlWithXML() {
+		WText text = new WText(TransformXMLTestHelper.TEST_XML);
+		text.setEncodeText(false);
+		// Dont use PageShell as it wraps the XML with ui:root and test xslt does not pass the other tags
+		String output = WebUtilities.renderWithTransformToHTML(new MockRequest(), text, false);
+		Assert.assertEquals("Invalid html output with XML", TransformXMLTestHelper.EXPECTED, output);
 	}
 
 //	/**